classifica pfpj e valida coluna
